refactor: extract BaseUnnamedFieldAggregator and update UnnamedFieldAggregator
Moved core logic from UnnamedFieldAggregator to new BaseUnnamedFieldAggregator for better reusability. Updated UnnamedFieldAggregator to inherit from the new base class and replaced detailed comments with <inheritdoc/>.
